SGX: UDD: Create sysfs entry to allow ignoring the display sync
This patch creates a sysfs entry that allows to tell the UDD
to ignore the synchronization with a specific display. The change
can be made on the fly without rebooting the kernel. This feature
makes sense only if flipping is enabled on the display. The
sysfs entry can be found in the directory

/sys/devices/platform/omaplfb/displayX/ignore_sync

By echoing a 1 to the ignore_sync entry the frames per second
will go beyond 60fps since the synchronization with the display
will be ignored. Setting it to 0 will return it to normal.

Signed-off-by: Gustavo Diaz Prado <x0083741@ti.com>
Change-Id: I242d88a9be202a7079c0ff7901751d4d996ef5be